--- packages/material-composer/src/lib/patchMaterial.ts
@@ -1,5 +1,0 @@
-import { Material } from "three"
-
-export const patchMaterial = (material: Material) => {
-  return material
-}
