+++ packages/patch-material/package.json
@@ -37,5 +37,8 @@
         }
       ]
     ]
+  },
+  "dependencies": {
+    "fp-ts": "^2.12.2"
   }
 }
